ssl fixes

+++ app/classService/tasks.py
@@ -2,7 +2,6 @@ from app import cel
 
 @cel.task
 def extract_quiz_task(filename):
-    print(filename)
     import zipfile
     zip_ref = zipfile.ZipFile(filename, 'r')
     zip_ref.extractall(filename[:-4])
@@ -12,10 +11,13 @@ def extract_quiz_task(filename):
 
 @cel.task
 def delete_quiz_resource(path):
+    import os
     import shutil
-    shutil.rmtree(path)
+    if os.path.exists(path):
+        shutil.rmtree(path)
 
 @cel.task
 def delete_resource(filename):
     import os
-    os.remove(filename)
+    if os.path.exists(filename):
+        os.remove(filename)
